server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
python-multipart>=0.0.12
httpx>=0.27.0
soundfile>=0.12.0
